refactor: Address AI review feedback on domain layer
- Change MeterReading, TariffStep, and Tariff from double to decimal
  for financial precision
- Remove unnecessary OrderBy in BillingCalculator.CalculateTieredCost
  (Tariff.Create already orders steps)
- Add null guard for Tariff.Create steps parameter
- Add null tariff tests for CalculateElectricityCost and
  CalculateTieredCost
- Rename misleading test to CalculateElectricityCost_WithValidTariff_DoesNotThrow
- Add missing sanitation tier 2/3 negative rate tests in PropertyTests
- Add test for null steps in Tariff.Create

+++ src/FlatRate.Domain/ValueObjects/MeterReading.cs
@@ -2,20 +2,21 @@ namespace FlatRate.Domain.ValueObjects;
 
 /// <summary>
 /// Represents a meter reading with opening and closing values.
+/// Uses decimal for precision in financial calculations.
 /// </summary>
 public sealed record MeterReading
 {
-    public double Opening { get; }
-    public double Closing { get; }
-    public double UnitsUsed => Closing - Opening;
+    public decimal Opening { get; }
+    public decimal Closing { get; }
+    public decimal UnitsUsed => Closing - Opening;
 
-    private MeterReading(double opening, double closing)
+    private MeterReading(decimal opening, decimal closing)
     {
         Opening = opening;
         Closing = closing;
     }
 
-    public static MeterReading Create(double opening, double closing)
+    public static MeterReading Create(decimal opening, decimal closing)
     {
         if (opening < 0)
             throw new ArgumentException("Opening reading cannot be negative.", nameof(opening));
