+++ src/FlatRate.Domain/ValueObjects/TariffStep.cs
@@ -2,26 +2,27 @@ namespace FlatRate.Domain.ValueObjects;
 
 /// <summary>
 /// Represents a single tier/step in a tariff structure.
+/// Uses decimal for precision in financial calculations.
 /// </summary>
 public sealed record TariffStep
 {
     /// <summary>
     /// The upper limit of units for this tier (inclusive).
     /// </summary>
-    public double UpperLimit { get; }
+    public decimal UpperLimit { get; }
 
     /// <summary>
     /// The rate per unit for this tier.
     /// </summary>
     public decimal Rate { get; }
 
-    private TariffStep(double upperLimit, decimal rate)
+    private TariffStep(decimal upperLimit, decimal rate)
     {
         UpperLimit = upperLimit;
         Rate = rate;
     }
 
-    public static TariffStep Create(double upperLimit, decimal rate)
+    public static TariffStep Create(decimal upperLimit, decimal rate)
     {
         if (upperLimit <= 0)
             throw new ArgumentException("Upper limit must be greater than zero.", nameof(upperLimit));
